resolve git error
